+++ edbee-lib/edbee/texteditorwidget.cpp
@@ -53,6 +53,7 @@ TextEditorWidget::TextEditorWidget( QWidget* parent)
     , scrollAreaRef_(0)
     , editCompRef_(0)
     , autoCompleteCompRef_(0)
+	, autoScrollMargin_(50)
 {
     // auto initialize edbee if this hasn't been done alread
     Edbee::instance()->autoInit();
@@ -118,7 +119,7 @@ TextEditorWidget::~TextEditorWidget()
 /// @param yPosIn the position in text-editor 'coordinates'
 void TextEditorWidget::scrollPositionVisible(int xPosIn, int yPosIn)
 {
-    scrollAreaRef_->ensureVisible( xPosIn, yPosIn );
+    scrollAreaRef_->ensureVisible(xPosIn, yPosIn, autoScrollMargin_, autoScrollMargin_);
 }
 
 
@@ -261,6 +262,17 @@ void TextEditorWidget::resizeEvent(QResizeEvent* event)
     updateRendererViewport();
 }
 
+/// Returns the auto scroll margin
+int TextEditorWidget::autoScrollMargin() const
+{
+    return autoScrollMargin_;
+}
+
+ /// Sets the auto scrollmargin
+void TextEditorWidget::setAutoScrollMargin(int amount)
+{
+    autoScrollMargin_ = amount;
+}
 
 /// a basic event-filter for recieving focus-events of the editor
 /// @param obj the object to filter the events for